Cleaned up utils - Moved main game features outside of 'utils' folder - Updated namespaces - Updated include paths
