+++ .vscode/tasks.json
@@ -9,28 +9,27 @@
                 "-g",
                 "${workspaceFolder}\\src\\*.cpp",
                 "${workspaceFolder}\\src\\battle\\*.cpp",
-                "${workspaceFolder}\\src\\creature\\*.cpp",
                 "${workspaceFolder}\\src\\create_character\\*.cpp",
+                "${workspaceFolder}\\src\\creature\\*.cpp",
                 "${workspaceFolder}\\src\\creature\\main_character\\*.cpp",
                 "${workspaceFolder}\\src\\creature\\main_character\\helpers\\*.cpp",
                 "${workspaceFolder}\\src\\creature\\main_character\\main_character_race_class\\*.cpp",
                 "${workspaceFolder}\\src\\loading_screen\\*.cpp",
+                "${workspaceFolder}\\src\\read_from_file\\*.cpp",
+                "${workspaceFolder}\\src\\save_and_load_game\\load_game\\*.cpp",
+                "${workspaceFolder}\\src\\save_and_load_game\\save_game\\*.cpp",
                 "${workspaceFolder}\\src\\start_menu\\*.cpp",
                 "${workspaceFolder}\\src\\start_menu\\continue_game\\*.cpp",
                 "${workspaceFolder}\\src\\start_menu\\main_settings\\*.cpp",
                 "${workspaceFolder}\\src\\start_menu\\start_game\\*.cpp",
                 "${workspaceFolder}\\src\\story\\*.cpp",
-                "${workspaceFolder}\\src\\utils\\clear_screen\\*.cpp",
                 "${workspaceFolder}\\src\\utils\\customize_text\\*.cpp",
+                "${workspaceFolder}\\src\\utils\\handle_sound\\*.cpp",
+                "${workspaceFolder}\\src\\utils\\print_and_clear_screen\\*.cpp",
+                "${workspaceFolder}\\src\\utils\\roll_dice\\*.cpp",
                 "${workspaceFolder}\\src\\utils\\user_choice\\*.cpp",
                 "${workspaceFolder}\\src\\utils\\user_choice\\helpers\\*.cpp",
                 "${workspaceFolder}\\src\\utils\\user_choice\\reserved_words\\*.cpp",
-                "${workspaceFolder}\\src\\utils\\handle_sound\\*.cpp",
-                "${workspaceFolder}\\src\\utils\\load_game\\*.cpp",
-                "${workspaceFolder}\\src\\utils\\print\\*.cpp",
-                "${workspaceFolder}\\src\\utils\\read_from_file\\*.cpp",
-                "${workspaceFolder}\\src\\utils\\roll_dice\\*.cpp",
-                "${workspaceFolder}\\src\\utils\\save_game\\*.cpp",
                 "-lWinmm",
                 "-o",
                 "${workspaceFolder}\\build\\text-adventure.exe"
