+++ src/main.cpp
@@ -1,9 +1,9 @@
-#include "./utils/utils.h"
-#include "./creature/main_character/main_character_prototypes.h"
-#include "./loading_screen/loading_screen.h"
-#include "./start_menu/start_menu.h"
-#include "./story/prologue.h"
-#include <iostream>
+#include "utils/utils.h"
+#include "creature/main_character/main_character_prototypes.h"
+#include "loading_screen/loading_screen.h"
+#include "save_and_load_game/load_game/load_game.h"
+#include "start_menu/start_menu.h"
+
 #include <time.h>
 
 int main()
@@ -15,7 +15,7 @@ int main()
 	MainCharacterUniquePtr main_character_unique_ptr =
 		std::make_unique<MainCharacter>("Player");
 
-	utils::LoadGame(*main_character_unique_ptr);
+	LoadGame(*main_character_unique_ptr);
 
 	StartMenu(*main_character_unique_ptr);
