+++ src/save_and_load_game/save_game/save_game.h
@@ -1,12 +1,8 @@
 #pragma once
 
 #include "../../creature/main_character/main_character.h"
-#include <iostream>
 
-namespace utils
-{
 /**
  * @brief A function that saves character stats in a txt document
  */
-void LoadGame(MainCharacter& main_character);
-} // namespace utils
+void SaveGame(MainCharacter& main_character);
