+++ src/start_menu/continue_game/continue_game.cpp
@@ -1,5 +1,6 @@
 #include "continue_game.h"
 #include "../../utils/utils.h"
+#include "../../save_and_load_game/load_game/load_game.h"
 #include "../../story/prologue.h"
 
 void ContinueGame(MainCharacter& main_character)
@@ -8,7 +9,7 @@ void ContinueGame(MainCharacter& main_character)
 
 	utils::ClearScreen();
 
-	utils::LoadGame(main_character);
+	LoadGame(main_character);
 
 	Prologue(main_character);
 }
