+++ src/utils/print_and_clear_screen/clear_screen.h
